Significantly modifies Main Activity file to better work with new layouts/fragments
New fragments are not yet added to layout.
However, main activity fragment is updated and appears to be functioning correctly.

Toolbar needs to be updated to allow switching of fragments, similar to as shown in the mockup.
However, fragments should probably be implimented first beforehand. Not sure exactly how to hook up multiple fragments to launch from one activity, as appears to be what we need? Original WICED code does not seem to follow the same principles we learned (and honestly looks kind of messy in how they impliment fragments) so referencing that isn't particularly helpful.