Reverts many changes from original development
Was under the impression that we wanted to essentially "gut" the program (as much as possible) and re-create it using concepts learned in class. IE using support fragments instead of built in, using abstraction to reduce redundancy, etc.
(Hopefully?) reverted back to Broadcom's original code, only better organized and with database branch merged in.

New layouts from original development are saved in a seperate "NewLayouts" folder within the standard layouts folder.